changed font, added art, added new game screen
there might be more idk but like thats about it
