--- Assets/User Folders/Jack/Main Menu stuff/menuAssets/Main Menu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4c63d2dd18f14bd4c8e329501a2f1fdb
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
