--- Assets/User Folders/Jack/Main Menu stuff/menuAssets/Options Menu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b8bfe11d043c0bf47908ec66bf81d997
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
