+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/Sprites slash art.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 56186f26e6f236748b0bd03b14461ff6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
