+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/asset thingys.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a067e4028fed9224687942c4483625b1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
