+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/asset thingys/pauseMenu.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6737b695c22fcff42b34e23672a74c91
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
