+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/font.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd901bc7a45acbd4b8bf23e0c4c079da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
