+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/font/Pixel Georgia Bold SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 852099fbdfd3c2f48bc2962ef14c442f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
