+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/font/Pixel Georgia Bold.ttf.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: b3584b380614ad142857d382b4188140
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Pixel Georgia?Bold
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
