+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/font/Pixel Georgia SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ae4a21ef281d9c429060ad7cc13de27
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
