+++ Assets/User Folders/Jack/Main Menu stuff/menuAssets/font/Pixel Georgia.ttf.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: d1de172ef404f5b43bc3031ef9f6969a
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Pixel Georgia
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
