--- Assets/User Folders/Jack/Main Menu stuff/menuAssets/pauseMenu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6737b695c22fcff42b34e23672a74c91
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
