--- Assets/User Folders/Jack/Scripts/Main Menu/AudioManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3646dd70660f863468114c5cd099530d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
