--- Assets/User Folders/Jack/Scripts/Main Menu/ManagerSingleton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f2622aeabbf819e40946c861b3d720a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
