--- Assets/User Folders/Jack/Scripts/Main Menu/Sound.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 56134e0a3efde104685f7f41eebec8fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
