+++ Assets/User Folders/Jack/Scripts/Main Menu/actual scripts/ColourPicker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7dc73a70cb61ecb47a223e291ae9b5ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
