+++ Assets/User Folders/Jack/Scripts/Main Menu/rand stuff idk what it does.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf568aae940da374e80070c0d1340d30
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
