+++ Assets/User Folders/Jack/Scripts/Main Menu/rand stuff idk what it does/scene Loader.cs
@@ -1,0 +1,29 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class sceneLoader : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void LoadScene1()
+    {
+        SceneManager.LoadScene("Scene1");
+    }
+
+    public void LoadScene2()
+    {
+        SceneManager.LoadScene("Scene2");
+    }
+}
