--- Assets/User Folders/Jack/Scripts/Main Menu/scene Loader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 82a97260ea7908c42b6e1ca51131a2c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
